requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
flask-cors==4.0.0
requests==2.31.0

# ML Dependencies
numpy>=1.21.0
Pillow>=8.0.0
scikit-learn>=1.0.0 